Update wallpaper method and reorder
I think setting the wallpaper was affecting the dock updates in a race condition situation, so now it updates the dock first. The wallpaper change method has been updated to use automator which makes the code clean and seems to work well. Suprising that I couldn't find a better way to do that, but it's an improvement.